requirements.txt
ecdsa==0.19.0
pycryptodome==3.20.0
sha3==0.2.1